pyproject.toml
[project]
name = "nyxcore"
version = "0.2.0"
description = "Local-first music library review and cleanup toolkit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.0.0",
  "mutagen>=1.47.0",
  "openai>=1.0.0",
  "aiohttp>=3.9.0",
  "sqlmodel>=0.0.22",
  "pydantic>=2.0.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
clap = [
  "torch",
  "torchaudio",
  "laion-clap",
]
audio-analysis = [
  "essentia",
]
web = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
]
dev = [
  "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
[tool.setuptools.packages.find]
include = ["nyxcore*"]

[tool.setuptools.package-data]
nyxcore = ["resources/*.yaml"]
